+++ CHANGELOG.md
@@ -49,7 +49,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
-- Add `MessageListController.hideUnreadLabel()` method to manually hide the unread separator. [#6001](https://github.com/GetStream/stream-chat-android/pull/6001)
+- Add `MessageListController.hideUnreadSeparator()` method to manually hide the unread separator. [#6001](https://github.com/GetStream/stream-chat-android/pull/6001)
 
 ### ⚠️ Changed
 
@@ -61,7 +61,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
-- Add `MessageListViewModel.hideUnreadLabel()` method to manually hide the unread separator. [#6001](https://github.com/GetStream/stream-chat-android/pull/6001)
+- Add `MessageListViewModel.hideUnreadSeparator()` method to manually hide the unread separator. [#6001](https://github.com/GetStream/stream-chat-android/pull/6001)
 
 ### ⚠️ Changed
 
@@ -73,7 +73,7 @@
 ### ⬆️ Improved
 
 ### ✅ Added
-- Add `MessageListViewModel.hideUnreadLabel()` method to manually hide the unread separator. [#6001](https://github.com/GetStream/stream-chat-android/pull/6001)
+- Add `MessageListViewModel.hideUnreadSeparator()` method to manually hide the unread separator. [#6001](https://github.com/GetStream/stream-chat-android/pull/6001)
 
 ### ⚠️ Changed